backend/requirements.txt
Flask
Flask-Cors
joblib
numpy
opencv-python-headless
lightgbm
scikit-learn
Pillow